Don't overwrite secret file if atomic temp file not written

+++ pkg/atomicwriter/atomic_writer.go
@@ -15,15 +15,25 @@ type osFuncs struct {
 	chmod    func(string, os.FileMode) error
 	rename   func(string, string) error
 	remove   func(string) error
+	sync     func(*os.File) error
+	tempFile func(string, string) (*os.File, error)
 	truncate func(string, int64) error
+	write    func(*os.File, []byte) (int, error)
 }
 
 // Instantiation of OS Function table using std OS
 var stdOSFuncs = osFuncs{
-	chmod:    os.Chmod,
-	rename:   os.Rename,
-	remove:   os.Remove,
+	chmod:  os.Chmod,
+	rename: os.Rename,
+	remove: os.Remove,
+	sync: func(file *os.File) error {
+		return file.Sync()
+	},
+	tempFile: ioutil.TempFile,
 	truncate: os.Truncate,
+	write: func(file *os.File, content []byte) (int, error) {
+		return file.Write(content)
+	},
 }
 
 type atomicWriter struct {
@@ -33,43 +43,55 @@ type atomicWriter struct {
 	os          osFuncs
 }
 
-// This package provides a simple atomic file writer which implements the
-// io.WriteCloser interface. This allows us to use AtomicWriter the way we
+// NewAtomicWriter provides a simple atomic file writer which implements the
+// io.WriteCloser interface. This allows us to use the atomic writer the way we
 // would use any other Writer, such as a Buffer. Additonally, this struct
 // takes the file path during construction, so the code which calls
 // `Write()` doesn't need to be concerned with the destination, just like
 // any other writer.
-func NewAtomicWriter(path string, permissions os.FileMode) (io.WriteCloser, error) {
+func NewAtomicWriter(path string, permissions os.FileMode) io.WriteCloser {
 	return newAtomicWriter(path, permissions, stdOSFuncs)
 }
 
-func newAtomicWriter(path string, permissions os.FileMode, osFuncs osFuncs) (io.WriteCloser, error) {
-	dir, file := filepath.Split(path)
-
-	f, err := ioutil.TempFile(dir, file)
-	if err != nil {
-		log.Error(messages.CSPFK055E, path)
-		return nil, err
-	}
-
+func newAtomicWriter(path string, permissions os.FileMode, osFuncs osFuncs) io.WriteCloser {
 	return &atomicWriter{
 		path:        path,
-		tempFile:    f,
+		tempFile:    nil,
 		permissions: permissions,
 		os:          osFuncs,
-	}, nil
+	}
 }
 
 func (w *atomicWriter) Write(content []byte) (n int, err error) {
+	// Create a temporary file if not created already
+	if w.tempFile == nil {
+		dir, file := filepath.Split(w.path)
+
+		f, err := w.os.tempFile(dir, file)
+		if err != nil {
+			log.Error(messages.CSPFK055E, w.path)
+			return 0, err
+		}
+		w.tempFile = f
+	}
+
 	// Write to the temporary file
-	return w.tempFile.Write(content)
+	n, err = w.os.write(w.tempFile, content)
+	if err != nil {
+		log.Error(messages.CSPFK061E, w.path)
+		w.Cleanup()
+	}
+	return n, err
 }
 
 func (w *atomicWriter) Close() error {
+	if w.tempFile == nil {
+		return nil
+	}
 	defer w.Cleanup()
 
 	// Flush and close the temporary file
-	err := w.tempFile.Sync()
+	err := w.os.sync(w.tempFile)
 	if err != nil {
 		log.Error(messages.CSPFK056E, w.tempFile.Name())
 		return err
@@ -89,6 +111,7 @@ func (w *atomicWriter) Close() error {
 		log.Error(messages.CSPFK058E, w.tempFile.Name(), w.path)
 		return err
 	}
+	w.tempFile = nil
 
 	return nil
 }
@@ -97,18 +120,26 @@ func (w *atomicWriter) Close() error {
 // the `Close()` method, but can also be called manually in cases where `Close()`
 // is not called.
 func (w *atomicWriter) Cleanup() {
+	if w.tempFile == nil {
+		return
+	}
+
 	err := w.os.remove(w.tempFile.Name())
-	if err == nil {
+	if err == nil || os.IsNotExist(err) {
+		w.tempFile = nil
 		return
 	}
 
 	// If we can't remove the temporary directory, truncate the file to remove all secret content
 	err = w.os.truncate(w.tempFile.Name(), 0)
-	if err == nil || os.IsNotExist(err) {
+	switch {
+	case os.IsNotExist(err):
+		// This shouldn't happen, but just to be safe.
+		w.tempFile = nil
+	case err != nil:
+		// Truncate failed as well
+		log.Error(messages.CSPFK060E, w.tempFile.Name(), w.path)
+	default:
 		log.Error(messages.CSPFK059E, w.tempFile.Name(), w.path)
-		return
 	}
-
-	// If that failed as well, log the error
-	log.Error(messages.CSPFK060E, w.tempFile.Name(), w.path)
 }
